# backend/openapi.yaml
openapi: 3.0.3
info:
  title: EduVerify API
  version: "1.0.0"
  description: >
    API REST del backend de EduVerify. Todas las respuestas siguen el envelope
    { status: 'success', data } / { status: 'error', message }.
    Ver backend/README.md para setup y arquitectura.
servers:
  - url: http://localhost:3001
security: []
...
paths:
...
  /api/playlists:
    get:
      tags: [playlists]
      summary: Listar playlists propias con sus videos
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Playlist' }
        "401": { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [playlists]
      summary: Crear playlist
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre]
              properties:
                nombre: { type: string, minLength: 1, maxLength: 255 }
      responses:
        "201":
          description: Creada
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Playlist' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /api/playlists/{id}:
    delete:
      tags: [playlists]
      summary: Eliminar playlist (solo el dueño)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessMessage' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
  /api/playlists/{id}/videos/{videoId}:
    post:
      tags: [playlists]
      summary: Añadir video a playlist (upsert, solo el dueño)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: videoId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessMessage' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [playlists]
      summary: Quitar video de playlist (solo el dueño)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: videoId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessMessage' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }

  # ── PROFESOR PLAYLISTS ───────────────────────────────────────────────────
  /api/profesor/playlists:
    get:
      tags: [profesor-playlists]
      summary: Listar playlists propias de profesor/creador con sus videos
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ProfesorPlaylist' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [profesor-playlists]
      summary: Crear playlist de profesor/creador
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre]
              properties:
                nombre: { type: string, minLength: 1, maxLength: 255 }
      responses:
        "201":
          description: Creada
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/ProfesorPlaylist' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /api/profesor/playlists/public/{userId}:
    get:
      tags: [profesor-playlists]
      summary: Listar playlists públicas de un profesor/creador
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ProfesorPlaylist' }
  /api/profesor/playlists/{id}:
    patch:
      tags: [profesor-playlists]
      summary: Renombrar playlist (solo el dueño)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre]
              properties:
                nombre: { type: string, minLength: 1, maxLength: 255 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/ProfesorPlaylist' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [profesor-playlists]
      summary: Eliminar playlist (solo el dueño)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessMessage' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
  /api/profesor/playlists/{id}/videos/{videoId}:
    post:
      tags: [profesor-playlists]
      summary: Añadir/reordenar video en playlist (upsert `orden`, solo el dueño)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: videoId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                orden: { type: integer, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessMessage' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [profesor-playlists]
      summary: Quitar video de playlist (solo el dueño)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: videoId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessMessage' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }

  # ── SUBSCRIPTIONS ─────────────────────────────────────────────────────────
  /api/subscriptions:
    get:
      tags: [subscriptions]
      summary: Listar profesores a los que el usuario está suscrito
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        professor_id: { type: integer }
                        nombre: { type: string }
                        notificaciones: { type: boolean }
                        avatar_url: { type: string, format: uri, nullable: true }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /api/subscriptions/{professorId}:
    post:
      tags: [subscriptions]
      summary: Suscribirse a un profesor (upsert, crea notificación)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: professorId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                notificaciones: { type: boolean, default: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessMessage' }
        "400":
          description: Intento de auto-suscripción
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
    delete:
      tags: [subscriptions]
      summary: Cancelar suscripción
      security: [{ bearerAuth: [] }]
      parameters:
        - name: professorId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessMessage' }
        "401": { $ref: '#/components/responses/Unauthorized' }

  # ── NOTIFICATIONS ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Token obtenido de /api/auth/login, /registro o /google. Payload:
        { sub: userId, email, rol, premium }. Expira en 7 días.
...
  responses:
    BadRequest:
      description: Body/params inválidos (Zod)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Token ausente, inválido o expirado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Rol insuficiente o no es el dueño del recurso
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    Categoria:
      type: string
      enum: [Programación, Ciberseguridad, Matemáticas, Electrónica, Arte]
    ErrorResponse:
      type: object
      properties:
        status: { type: string, example: error }
        message: { type: string }
    SuccessMessage:
      type: object
      properties:
        status: { type: string, example: success }
        data:
          type: object
          properties:
            message: { type: string }
...
    Video:
      type: object
      properties:
        id: { type: integer }
        titulo: { type: string }
        descripcion: { type: string, nullable: true }
        url_video: { type: string, format: uri }
        categoria: { $ref: '#/components/schemas/Categoria' }
        tipo: { type: string, enum: [grabado, envivo] }
        es_premium: { type: boolean }
        vistas: { type: integer }
        duracion: { type: string }
        created_at: { type: string, format: date-time }
        autor: { type: string }
        autor_id: { type: integer }
        author_avatar_url: { type: string, format: uri, nullable: true }
...
    Playlist:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer }
        nombre: { type: string }
        created_at: { type: string, format: date-time }
        videos:
          type: array
          items: { $ref: '#/components/schemas/Video' }
    ProfesorPlaylist:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer }
        nombre: { type: string }
        created_at: { type: string, format: date-time }
        videos:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/Video'
              - type: object
                properties:
                  orden: { type: integer }
...